requirements.txt
pymongo==4.6.1
python-dotenv==1.0.0
PyPDF2==3.0.1
google-generativeai==0.3.1